03/RAG_projcet/requirements.txt
llama-index
llama-index-llms-groq
llama-index-embeddings-cohere
llama-index-core
gradio
python-dotenv
pydantic
nest_asyncio